refactor schedule context
